requirements.txt
implicit
pandas
pandas-gbq
scipy
python-dotenv
numpy
Flask
Flask-BasicAuth
gunicorn
